re-org flags code
