+++ cmd/init.go
@@ -11,6 +11,8 @@ import (
 	"github.com/kubefirst/kubefirst/configs"
 	"github.com/kubefirst/kubefirst/internal/aws"
 	"github.com/kubefirst/kubefirst/internal/downloadManager"
+	"github.com/kubefirst/kubefirst/internal/flagset"
+	"github.com/kubefirst/kubefirst/internal/progressPrinter"
 	"github.com/kubefirst/kubefirst/internal/telemetry"
 	"github.com/kubefirst/kubefirst/pkg"
 	"github.com/spf13/cobra"
@@ -27,53 +29,47 @@ and usage of using your command. For example:
 Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-
+	RunE: func(cmd *cobra.Command, args []string) error {
+		progressPrinter.GetInstance()
+		progressPrinter.SetupProgress(10)
 		config := configs.ReadConfig()
 
-		dryRun, err := cmd.Flags().GetBool("dry-run")
+		globalFlags, err := flagset.ProcessGlobalFlags(cmd)
 		if err != nil {
-			log.Panic(err)
+			return err
 		}
 
-		useTelemetry, err := cmd.Flags().GetBool("use-telemetry")
+		installerFlags, err := processInstallerGenericFlags(cmd)
 		if err != nil {
-			log.Panic(err)
+			return err
 		}
 
-		if !useTelemetry {
-			log.Println("telemetry is disabled")
-		}
-
-		log.Println("dry run enabled:", dryRun)
-
-		arnRole, err := cmd.Flags().GetString("aws-assume-role")
+		awsFlags, err := flagset.ProcessAwsFlags(cmd)
 		if err != nil {
-			log.Println("unable to use the provided AWS IAM role for AssumeRole feature")
-			return
+			return err
 		}
 
-		if len(arnRole) > 0 {
+		log.Println("dry run enabled:", globalFlags.DryRun)
+
+		if len(awsFlags.AssumeRole) > 0 {
 			log.Println("calling assume role")
-			err := aws.AssumeRole(arnRole)
+			err := aws.AssumeRole(awsFlags.AssumeRole)
 			if err != nil {
 				log.Println(err)
-				return
+				return err
 			}
-			log.Printf("assuming new AWS credentials based on role %q", arnRole)
+			log.Printf("assuming new AWS credentials based on role %q", awsFlags.AssumeRole)
 		}
 
-		pkg.SetupProgress(10)
-		trackers := pkg.GetTrackers()
-		trackers[pkg.DownloadDependencies] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.DownloadDependencies, 3)}
-		trackers[pkg.GetAccountInfo] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.GetAccountInfo, 1)}
-		trackers[pkg.GetDNSInfo] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.GetDNSInfo, 1)}
-		trackers[pkg.TestHostedZoneLiveness] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.TestHostedZoneLiveness, 1)}
-		trackers[pkg.CloneAndDetokenizeGitOpsTemplate] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.CloneAndDetokenizeGitOpsTemplate, 1)}
-		trackers[pkg.CloneAndDetokenizeMetaphorTemplate] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.CloneAndDetokenizeMetaphorTemplate, 1)}
-		trackers[pkg.CreateSSHKey] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.CreateSSHKey, 1)}
-		trackers[pkg.CreateBuckets] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.CreateBuckets, 1)}
-		trackers[pkg.SendTelemetry] = &pkg.ActionTracker{Tracker: pkg.CreateTracker(pkg.SendTelemetry, 1)}
+		progressPrinter.AddTracker("step-download", pkg.DownloadDependencies, 3)
+		progressPrinter.AddTracker("step-account", pkg.GetAccountInfo, 1)
+		progressPrinter.AddTracker("step-dns", pkg.GetDNSInfo, 1)
+		progressPrinter.AddTracker("step-live", pkg.TestHostedZoneLiveness, 1)
+		progressPrinter.AddTracker("step-gitops", pkg.CloneAndDetokenizeGitOpsTemplate, 1)
+		progressPrinter.AddTracker("step-metaphor", pkg.CloneAndDetokenizeMetaphorTemplate, 1)
+		progressPrinter.AddTracker("step-ssh", pkg.CreateSSHKey, 1)
+		progressPrinter.AddTracker("step-buckets", pkg.CreateBuckets, 1)
+		progressPrinter.AddTracker("step-telemetry", pkg.SendTelemetry, 1)
 
 		k1Dir := fmt.Sprintf("%s", config.K1FolderPath)
 		if _, err := os.Stat(k1Dir); errors.Is(err, os.ErrNotExist) {
@@ -89,8 +85,8 @@ to quickly create a Cobra application.`,
 		metricName := "kubefirst.init.started"
 		metricDomain := hostedZoneName
 
-		if !dryRun {
-			telemetry.SendTelemetry(useTelemetry, metricDomain, metricName)
+		if !globalFlags.DryRun {
+			telemetry.SendTelemetry(globalFlags.UseTelemetry, metricDomain, metricName)
 		} else {
 			log.Printf("[#99] Dry-run mode, telemetry skipped:  %s", metricName)
 		}
@@ -108,85 +104,27 @@ to quickly create a Cobra application.`,
 		viper.Set("argocd.local.service", "http://localhost:8080")
 		viper.Set("gitlab.local.service", "http://localhost:8888")
 		viper.Set("vault.local.service", "http://localhost:8200")
-		// admin email
 		// used for letsencrypt notifications and the gitlab root account
-		adminEmail, _ := cmd.Flags().GetString("admin-email")
-		log.Println("adminEmail:", adminEmail)
-		viper.Set("adminemail", adminEmail)
 
-		// set region
-		region, err := cmd.Flags().GetString("region")
-		if err != nil {
-			log.Panicf("unable to get region values from viper")
-		}
-		viper.Set("aws.region", region)
-		// propagate it to local environment
-		err = os.Setenv("AWS_REGION", region)
-		if err != nil {
-			log.Panicf("unable to set environment variable AWS_REGION, error is: %v", err)
-		}
-		log.Println("region:", region)
-
-		// set profile
-		profile, err := cmd.Flags().GetString("profile")
-		if err != nil {
-			log.Panicf("unable to get region values from viper")
-		}
-		viper.Set("aws.profile", profile)
-		// propagate it to local environment
-		err = os.Setenv("AWS_PROFILE", profile)
-		if err != nil {
-			log.Panicf("unable to set environment variable AWS_PROFILE, error is: %v", err)
-		}
-		log.Println("profile:", profile)
-
-		nodes_spot, err := cmd.Flags().GetBool("aws-nodes-spot")
-		if err != nil {
-			log.Panic(err)
-		}
-		viper.Set("aws.nodes_spot", nodes_spot)
-		log.Println("aws.nodes_spot: ", nodes_spot)
-
-		// cluster name
-		clusterName, err := cmd.Flags().GetString("cluster-name")
-		if err != nil {
-			log.Panic(err)
-		}
-		viper.Set("cluster-name", clusterName)
-		log.Println("cluster-name:", clusterName)
-
-		// version-gitops
-		versionGitOps, err := cmd.Flags().GetString("version-gitops")
-		if err != nil {
-			log.Panic(err)
-		}
-		viper.Set("version-gitops", versionGitOps)
-		log.Println("version-gitops:", versionGitOps)
-
-		bucketRand, err := cmd.Flags().GetString("s3-suffix")
-		if err != nil {
-			log.Panic(err)
-		}
-		viper.Set("bucket.rand", bucketRand)
-		log.Println("s3-suffix:", clusterName)
+		log.Println("s3-suffix:", installerFlags.ClusterName)
 
 		viper.WriteConfig()
 
 		//! tracker 0
 		log.Println("installing kubefirst dependencies")
-		trackers[pkg.DownloadDependencies].Tracker.Increment(1)
-		err = downloadManager.DownloadTools(config, trackers)
+		progressPrinter.IncrementTracker("step-download", 1)
+		err = downloadManager.DownloadTools(config)
 		if err != nil {
-			log.Panic(err)
+			return err
 		}
 		log.Println("dependency installation complete")
-		trackers[pkg.DownloadDependencies].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-download", 2)
 
 		//! tracker 1
 		log.Println("getting aws account information")
 		aws.GetAccountInfo()
 		log.Printf("aws account id: %s\naws user arn: %s", viper.GetString("aws.accountid"), viper.GetString("aws.userarn"))
-		trackers[pkg.GetAccountInfo].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-account", 1)
 
 		//! tracker 2
 		// hosted zone id
@@ -194,32 +132,32 @@ to quickly create a Cobra application.`,
 		hostedZoneId := aws.GetDNSInfo(hostedZoneName)
 		// viper values set in above function
 		log.Println("hostedZoneId:", hostedZoneId)
-		trackers[pkg.GetDNSInfo].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-dns", 1)
 
 		//! tracker 3
 		// todo: this doesn't default to testing the dns check
 		skipHostedZoneCheck := viper.GetBool("init.hostedzonecheck.enabled")
 		if !skipHostedZoneCheck {
-			aws.TestHostedZoneLiveness(dryRun, hostedZoneName, hostedZoneId)
+			aws.TestHostedZoneLiveness(globalFlags.DryRun, hostedZoneName, hostedZoneId)
 		} else {
 			log.Println("skipping hosted zone check")
 		}
-		trackers[pkg.TestHostedZoneLiveness].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-live", 1)
 
 		//! tracker 4
 		//* should we consider going down to a single bucket
 		//* for state and artifacts on open source?
 		//* hitting a bucket limit on an install might deter someone
 		log.Println("creating buckets for state and artifacts")
-		aws.BucketRand(dryRun)
-		trackers[pkg.CreateBuckets].Tracker.Increment(1)
+		aws.BucketRand(globalFlags.DryRun)
+		progressPrinter.IncrementTracker("step-buckets", 1)
 		log.Println("BucketRand() complete")
 
 		//! tracker 5
 		log.Println("creating an ssh key pair for your new cloud infrastructure")
 		pkg.CreateSshKeyPair()
 		log.Println("ssh key pair creation complete")
-		trackers[pkg.CreateSSHKey].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-ssh", 1)
 
 		//! tracker 6
 		// TODO: get the below line added as a legit flag, don't merge with any value except kubefirst
@@ -230,18 +168,18 @@ to quickly create a Cobra application.`,
 		}
 		prepareKubefirstTemplateRepo(config, gitopsTemplateGithubOrgOverride, "gitops", viper.GetString("version-gitops"))
 		log.Println("clone and detokenization of gitops-template repository complete")
-		trackers[pkg.CloneAndDetokenizeGitOpsTemplate].Tracker.Increment(int64(1))
+		progressPrinter.IncrementTracker("step-gitops", 1)
 
 		//! tracker 7
 		log.Printf("cloning and detokenizing the metaphor-template repository")
 		prepareKubefirstTemplateRepo(config, "kubefirst", "metaphor", "main")
 		log.Println("clone and detokenization of metaphor-template repository complete")
-		trackers[pkg.CloneAndDetokenizeMetaphorTemplate].Tracker.Increment(int64(1))
+		progressPrinter.IncrementTracker("step-metaphor", 1)
 
 		metricName = "kubefirst.init.completed"
 
-		if !dryRun {
-			telemetry.SendTelemetry(useTelemetry, metricDomain, metricName)
+		if !globalFlags.DryRun {
+			telemetry.SendTelemetry(globalFlags.UseTelemetry, metricDomain, metricName)
 		} else {
 			log.Printf("[#99] Dry-run mode, telemetry skipped:  %s", metricName)
 		}
@@ -249,54 +187,26 @@ to quickly create a Cobra application.`,
 		viper.WriteConfig()
 
 		//! tracker 8
-		trackers[pkg.SendTelemetry].Tracker.Increment(1)
+		progressPrinter.IncrementTracker("step-telemetry", 1)
 		time.Sleep(time.Millisecond * 100)
+
+		return nil
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(initCmd)
-
-	initCmd.Flags().String("hosted-zone-name", "", "the domain to provision the kubefirst platform in")
-	err := initCmd.MarkFlagRequired("hosted-zone-name")
-	if err != nil {
-		log.Panic(err)
-	}
-	initCmd.Flags().String("admin-email", "", "the email address for the administrator as well as for lets-encrypt certificate emails")
-	err = initCmd.MarkFlagRequired("admin-email")
-	if err != nil {
-		log.Panic(err)
-	}
-	initCmd.Flags().String("cloud", "", "the cloud to provision infrastructure in")
-	err = initCmd.MarkFlagRequired("cloud")
-	if err != nil {
-		log.Panic(err)
-	}
-	initCmd.Flags().String("region", "eu-west-1", "the region to provision the cloud resources in")
-	err = initCmd.MarkFlagRequired("region")
-	if err != nil {
-		log.Panic(err)
-	}
-
-	initCmd.Flags().String("profile", "default", "AWS profile located at ~/.aws/config")
-	err = initCmd.MarkFlagRequired("profile")
-	if err != nil {
-		log.Panic(err)
-	}
-	initCmd.Flags().Bool("clean", false, "delete any local kubefirst content ~/.kubefirst, ~/.k1")
-
+	currentCommand := initCmd
+	log.Println("init started")
 	log.SetPrefix("LOG: ")
 	log.SetFlags(log.Ldate | log.Lmicroseconds | log.Llongfile)
 
-	initCmd.Flags().Bool("dry-run", false, "set to dry-run mode, no changes done on cloud provider selected")
-	log.Println("init started")
+	// Do we need this?
+	initCmd.Flags().Bool("clean", false, "delete any local kubefirst content ~/.kubefirst, ~/.k1")
 
-	initCmd.Flags().String("cluster-name", "kubefirst", "the cluster name, used to identify resources on cloud provider")
-	initCmd.Flags().String("s3-suffix", "", "unique identifier for s3 buckets")
-	initCmd.Flags().String("version-gitops", "main", "version/branch used on git clone")
-
-	// AWS assume role
-	initCmd.Flags().String("aws-assume-role", "", "instead of using AWS IAM user credentials, AWS AssumeRole feature generate role based credentials, more at https://docs.aws.amazon.com/STS/latest/APIReference/API_AssumeRole.html")
-	initCmd.Flags().Bool("use-telemetry", true, "installer will not send telemetry about this installation")
-	initCmd.Flags().Bool("aws-nodes-spot", false, "nodes spot on AWS EKS compute nodes")
+	//Group Flags
+	flagset.DefineGlobalFlags(currentCommand)
+	defineGithubCmdFlags(currentCommand)
+	flagset.DefineAWSFlags(currentCommand)
+	defineInstallerGenericFlags(currentCommand)
 }
